Move HipChat RequestDecoder to its own file and fix room ID bug
